+++ test/TestBankTest.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
 
-import {Test, console} from "forge-std/Test.sol";
-import {Bank} from "../src/Bank.sol";
+import { Test, console } from "forge-std/Test.sol";
+import { Bank } from "../src/Bank.sol";
 
 contract TestBankTest is Test {
     Bank public bank;
@@ -12,11 +12,11 @@ contract TestBankTest is Test {
     }
 
     function test_depositETH() public {
-        uint oldBanlance = address(this).balance; //当前钱包余额
-        uint depositAmount = 1 ether; //存款金额
+        uint256 oldBanlance = address(this).balance; //当前钱包余额
+        uint256 depositAmount = 1 ether; //存款金额
 
         vm.deal(address(this), depositAmount);
-        bank.depositETH{value: depositAmount}();
+        bank.depositETH{ value: depositAmount }();
 
         // vm.expectEmit(true, true, false, true);
         // emit bank.Deposit(address(this), depositAmount);
